rearrang files  to make adding graphics files more managable
