--- ChessPieces.h
@@ -1,19 +1,0 @@
-/*ChessPieces.h*/#pragma once
-
-enum Pieces : signed char
-{
-	empty = 0,
-	blackPawn = -1,
-	blackRook = -2,
-	blackKnight = -3,
-	blackBishop = -4,
-	blackQueen = -5,
-	blackKing = -6,
-	whitePawn = 1,
-	whiteRook = 2,
-	whiteKnight = 3,
-	whiteBishop = 4,
-	whiteQueen = 5,
-	whiteKing = 6,
-};
-
